third_party/blink/web_tests/platform/win/external/wpt/selection/selection-direction-on-double-click.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] direction returns "none" when there is a double click selection(directionless)
  assert_equals: expected "none" but got "forward"
Harness: the test ran to completion.
